Crate exar_server

Exar DB's server

This module contains a server implementation that uses Exar DB's TCP protocol.

It uses the one thread per connection model.

Server Initialization

extern crate exar;
extern crate exar_server;

use exar::*;
use exar_server::*;

let db = Database::new(DatabaseConfig::default());

let server_config = ServerConfig::default();
let server = Server::new(server_config.clone(), db).unwrap();

println!("ExarDB's server running at {}", server_config.address());
server.listen();
println!("ExarDB's server shutting down");

Structs

Credentials
A structure containing credentials.

Handler
Exar DB's server connection handler.

Server
Exar DB's server.

ServerConfig
Exar DB's server configuration.

Enums

ActionResult
A list specifying categories of connection handler action results.

State
A list specifying categories of connection state.